Add some Leo tests, update abi in frontend

+++ web/src/config/abi/ILeo.ts
@@ -27,8 +27,8 @@ export const ILeo = [
       },
       {
         name: "perSecond",
-        type: "uint256",
-        internalType: "uint256",
+        type: "uint64",
+        internalType: "uint64",
       },
       {
         name: "token",
@@ -196,8 +196,8 @@ export const ILeo = [
       },
       {
         name: "perSecond",
-        type: "uint256",
-        internalType: "uint256",
+        type: "uint64",
+        internalType: "uint64",
       },
       {
         name: "token",
@@ -286,8 +286,8 @@ export const ILeo = [
       },
       {
         name: "perSecond",
-        type: "uint256",
-        internalType: "uint256",
+        type: "uint64",
+        internalType: "uint64",
       },
       {
         name: "token",
